ReadFile first draft done.
Now uses smart pointers so I don't have to manully delete.
readFile now returns the parsed data, so I don't have to rely on modifying globals outside of a namespace, like in the HW2 skeleton.
Split scene shapes into separate Triangle and Sphere for now., to avoid using pointers to shapes to get the shape list to work.
TransformStack::rotate now asks for the angle in radians, to be consistent with the rest of the program.