First render of scene5.
Left ambient on the high debug setting in scene5.test (OOPS!), so the spheres are too white. Otherwise, the reflections look good and the highlights are the right size/in the right place.

It took a long time to run, so I will work on improving efficiency and perhaps do a release executable before running it again.