Build first release version. Re-render scene5
The render is miles faster with the release version! Only took a minute instead of an hour :smile: . However, towards the back of the image there is some loss of detail: some of the pixels are too bright/white. I will have to investigate that further, but for now, I want to focus on simplifying the code.